+++ lib/entities/coverage.dart
@@ -3,6 +3,7 @@ class Coverage {
   final double _minCoverage;
   double _totalCoverage = 0;
 
+  /// if [isCovered]
   bool get isCovered {
     if (_minCoverage <= 0) {
       return true;
@@ -10,12 +11,15 @@ class Coverage {
     return totalCoverage >= minCoverage;
   }
 
+  /// [minCoverage]
   double get minCoverage => double.parse(_minCoverage.toStringAsFixed(1));
 
+  /// [totalCoverage]
   double get totalCoverage => double.parse(_totalCoverage.toStringAsFixed(1));
   set totalCoverage(double value) {
     _totalCoverage = value;
   }
 
+  /// [Coverage] constructor
   Coverage(this._minCoverage);
 }